Rename utf8.{h,c} to unicode.{h,c}
