+++ src/c/src/unicode.c
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "utf8.h"
+#include "unicode.h"
 
 #include <string.h>
